Unit test: make flash mocking generic
Avoid duplicated test code for mocking the flash
